completed the first task
With the help of JS, I implemented the following actions:

When you first click on an element on the page, the background and text color changes using the getElementById() method, and when you click on the next element, the background and text color changes using the querySelector() method. When you click again on the corresponding elements, their background and text colors change.