completed the second task
Wrote a 6x6 table. When hovering over cell 3 in the table, the color of this cell changes to random. When you click on this cell, its color changes to the one specified by the user. When you double-click cell 3, the main diagonal is drawn in the color specified by the user.